requirements.txt
eth-testrpc
flask
py-solc
web3